Improved accessibility and stopped the user from compairing codes befor they were incrypted under certain surcumstances
Increased the delay for sleep in the clear function to 1 second so that the menu will be read by screen readers propperly as a smaller delay can sometimes make the menu unreadable.
Made it so that op3 and op4 are set to false in the enterNumbers function to stop the user from being able to compair a code befor incrypting it on the second time entering a code.