Added the messageCheckerBot, updated instructions and added run.bat files for both bots
Introduced a new bot that will time out a user who sends a message with a banned word in it.
The broadcaster can modify the program to add any words they wish.
Also updated the autoBanBot instructions to make them more user friendly and added instructions for the new bot.
Updated the readme to reflect these changes.